audio_router/foreground.py
import win32gui
import win32process
import psutil

def get_foreground_process():
    hwnd = win32gui.GetForegroundWindow()
    _, pid = win32process.GetWindowThreadProcessId(hwnd)
    return psutil.Process(pid).name()
