requirements.txt
pywin32
psutil
keyboard
pystray
pillow
customtkinter
